# adodasopenapi.yaml
openapi: 3.0.3
info:
  title: Adodas API
  description: |
    API REST completa para gestión de **clientes, ítems, pedidos, reembolsos y servicios**.
    
    ## ✨ Características:
    - ✅ **CRUD completo** todos los recursos
    - ✅ **Validaciones robustas** (DNI, emails, precios)
    - ✅ **Filtros avanzados** por estado, precio, cliente
    - ✅ **Errores específicos** 400/404/500 diferenciados
    - ✅ **Ejemplos reales** en todas las respuestas
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Desarrollo Local
...
components:
  schemas:
    ErrorResponseDto:
      type: object
      properties:
        code:
          type: integer
          description: Código HTTP del error
        message:
          type: string
          description: Descripción del error
      required: [ code, message ]
...
    ServiceDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 42
        name:
          type: string
          description: Nombre del servicio de la tienda.
          example: "Dobladillo de pantalones"
        description:
          type: string
          description: Descripción del servicio de ropa.
          example: "Ajuste profesional de dobladillo para pantalones vaqueros o vestir"
        price:
          type: number
          format: float
          minimum: 0
          description: Precio del servicio.
          example: 12.50
        availableFrom:
          type: string
          format: date
          description: Fecha desde la que está disponible.
          example: "2026-02-01"
        availableTo:
          type: string
          format: date
          description: Fecha hasta la que está disponible.
          example: "2026-12-31"
        isActive:
          type: boolean
          description: Indica si el servicio está disponible.
          example: true
        customerId:
          type: integer
          format: int64
          description: ID del cliente asociado (opcional).
          example: 15
        notes:
          type: string
          description: Notas adicionales del servicio.
          example: "Servicio exprés disponible (+5€)"
      required: [ name, price ]
...
paths:
...
  /services:
    get:
      tags: [ SERVICES ]
      summary: Lista todos los servicios
      description: Obtiene una lista de servicios con filtros opcionales por nombre, rango de precio y estado activo.
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filtra por nombre del servicio.
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: 0
          description: Precio mínimo de los servicios.
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
            format: float
            minimum: 0
          description: Precio máximo de los servicios.
        - name: isActive
          in: query
          required: false
          schema:
            type: boolean
          description: Filtra por servicios activos/inactivos.
      responses:
        '200':
          description: Lista de servicios obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceDto'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    post:
      tags: [ SERVICES ]
      summary: Crea un nuevo servicio
      description: Registra un nuevo servicio profesional validando nombre y precio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceDto'
      responses:
        '201':
          description: Servicio creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
  /services/{serviceId}:
    get:
      tags: [ SERVICES ]
      summary: Obtiene un servicio por su ID
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del servicio.
      responses:
        '200':
          description: Servicio encontrado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDto'
        '404':
          description: Servicio no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Servicio no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    put:
      tags: [ SERVICES ]
      summary: Modifica un servicio existente
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del servicio a modificar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceDto'
      responses:
        '200':
          description: Servicio modificado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Servicio no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Servicio no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    patch:
      tags: [ SERVICES ]
      summary: Actualiza parcialmente un servicio
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del servicio a actualizar parcialmente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                price: 1499.00
                isActive: false
                notes: "Servicio pausado temporalmente"
      responses:
        '200':
          description: Servicio actualizado parcialmente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDto'
        '400':
          description: Error de validación o cuerpo inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Servicio no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Servicio no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    delete:
      tags: [ SERVICES ]
      summary: Elimina un servicio
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del servicio a eliminar.
      responses:
        '204':
          description: Servicio eliminado correctamente (sin contenido en la respuesta).
        '404':
          description: Servicio no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Servicio no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"